+++ sjf4j-jdk17-test/src/test/java/org/sjf4j/jdk17/node/ReflectUtilTest.java
@@ -61,15 +61,9 @@ class ReflectUtilTest {
     // ---------- 2) getFieldName / getFieldAliases priority ----------
 
     @Test
-    void getFieldName_prefersNodeProperty_overOthers() throws Exception {
-        Field f = FieldNamePriorityPojo.class.getDeclaredField("f");
-        assertEquals("nodeName", ReflectUtil.getFieldName(f));
-    }
-
-    @Test
     void getFieldAliases_prefersNodePropertyAliases_overOthers() throws Exception {
         Field f = FieldNamePriorityPojo.class.getDeclaredField("f");
-        assertArrayEquals(new String[]{"na1", "na2"}, ReflectUtil.getFieldAliases(f));
+        assertArrayEquals(new String[]{"na1", "na2"}, ReflectUtil.getAliases(f));
     }
 
     static class FieldNamePriorityPojo {
@@ -83,28 +77,12 @@ class ReflectUtilTest {
     // ---------- 3) getParameterName / getParameterAliases ----------
 
     @Test
-    void getParameterName_prefersAnnotations_andReturnsNullWhenNoParamNames() throws Exception {
-        Constructor<CreatorCtorPojo> c = CreatorCtorPojo.class.getDeclaredConstructor(String.class, int.class);
-
-        // p0: has JsonProperty
-        assertEquals("name2", ReflectUtil.getParameterName(c.getParameters()[0]));
-
-        // p1: no any annotation
-        String p1 = ReflectUtil.getParameterName(c.getParameters()[1]);
-        if (!c.getParameters()[1].isNamePresent()) {
-            assertNull(p1);
-        } else {
-            assertEquals("age", p1);
-        }
-    }
-
-    @Test
     void getParameterAliases_readsAliasAnnotations() throws Exception {
         Constructor<CreatorCtorPojo> c = CreatorCtorPojo.class.getDeclaredConstructor(String.class, int.class);
 
-        assertArrayEquals(new String[]{"aliasN1", "aliasN2"}, ReflectUtil.getParameterAliases(c.getParameters()[0]));
+        assertArrayEquals(new String[]{"aliasN1", "aliasN2"}, ReflectUtil.getAliases(c.getParameters()[0]));
 
-        assertNull(ReflectUtil.getParameterAliases(c.getParameters()[1]));
+        assertNull(ReflectUtil.getAliases(c.getParameters()[1]));
     }
 
     static class CreatorCtorPojo {