+++ sjf4j/src/main/java/org/sjf4j/facade/fastjson2/Fastjson2StreamingIO.java
@@ -247,7 +247,7 @@ public class Fastjson2StreamingIO {
                 if (fi != null) {
                     Object vv = _readField(reader, fi, ownerType, ownerRawClazz);
                     fi.invokeSetterIfPresent(pojo, vv);
-                } else if (pi.isJojo) {
+                } else if (pi.isJojo && pi.readDynamic) {
                     if (dynamicMap == null) {
                         dynamicMap = new LinkedHashMap<>();
                     }
@@ -322,7 +322,7 @@ public class Fastjson2StreamingIO {
                 continue;
             }
 
-            if (pi.isJojo) {
+            if (pi.isJojo && pi.readDynamic) {
                 Object vv = _readRawNode(reader);
                 session.acceptResolvedJsonEntry(-1, key, vv);
                 if (parentAnyOfKey != null && parentAnyOfKey.equals(key)) {
@@ -604,6 +604,19 @@ public class Fastjson2StreamingIO {
             }
 
             if (node instanceof JsonObject) {
+                if (rawClazz != JsonObject.class) {
+                    NodeRegistry.PojoInfo pi = NodeRegistry.registerPojoOrElseThrow(rawClazz);
+                    if (!pi.writeDynamic) {
+                        writer.startObject();
+                        for (Map.Entry<String, NodeRegistry.FieldInfo> entry : pi.fields.entrySet()) {
+                            writer.writeName(entry.getKey());
+                            writer.writeColon();
+                            _writeNode(writer, entry.getValue().invokeGetter(node));
+                        }
+                        writer.endObject();
+                        return;
+                    }
+                }
                 writer.startObject();
                 ((JsonObject) node).forEach((k, v) -> {
                     try {