+++ sjf4j/src/test/java/org/sjf4j/node/ReflectUtilTest.java
@@ -58,12 +58,6 @@ class ReflectUtilTest {
     // ---------- 2) getFieldName / getFieldAliases priority ----------
 
     @Test
-    void getFieldName_prefersNodeProperty_overOthers() throws Exception {
-        Field f = FieldNamePriorityPojo.class.getDeclaredField("f");
-        assertEquals("nodeName", ReflectUtil.getFieldName(f, FieldNamePriorityPojo.class));
-    }
-
-    @Test
     void getFieldAliases_prefersNodePropertyAliases_overOthers() throws Exception {
         Field f = FieldNamePriorityPojo.class.getDeclaredField("f");
         assertArrayEquals(new String[]{"na1", "na2"}, ReflectUtil.getAliases(f));
@@ -78,22 +72,6 @@ class ReflectUtilTest {
     }
 
     // ---------- 3) getParameterName / getParameterAliases ----------
-
-    @Test
-    void getParameterName_prefersAnnotations_andReturnsNullWhenNoParamNames() throws Exception {
-        Constructor<CreatorCtorPojo> c = CreatorCtorPojo.class.getDeclaredConstructor(String.class, int.class);
-
-        // p0: has JsonProperty
-        assertEquals("name2", ReflectUtil.getParameterName(c.getParameters()[0], CreatorCtorPojo.class));
-
-        // p1: no any annotation
-        String p1 = ReflectUtil.getParameterName(c.getParameters()[1], CreatorCtorPojo.class);
-        if (!c.getParameters()[1].isNamePresent()) {
-            assertNull(p1);
-        } else {
-            assertEquals("age", p1);
-        }
-    }
 
     @Test
     void getParameterAliases_readsAliasAnnotations() throws Exception {